build warnings

+++ TouchChart/TouchChart/Classes/Models/SYSegment.m
@@ -726,7 +726,7 @@
     CGFloat angleDeg = [self angleDeg];
 
     // Si no esta ajustado responde que no
-    CGFloat resultAbs = fabsf(angleDeg);
+    CGFloat resultAbs = fabs(angleDeg);
     if (resultAbs > 180.0){
         // shouldn't this be resultAbs -= 180 ?
         resultAbs = resultAbs * 0.5;
